requirements.txt
flask
flask-cors
llama-cpp-python
sentence-transformers
scikit-learn
